app module refactor
