+++ .idea/androidTestResultsUserPreferences.xml
@@ -35,9 +35,11 @@
             <AndroidTestResultsTableState>
               <option name="preferredColumnWidths">
                 <map>
+                  <entry key="9d5e7a8" value="120" />
                   <entry key="Duration" value="90" />
                   <entry key="Pixel_6_API_33" value="120" />
                   <entry key="Tests" value="360" />
+                  <entry key="Xiaomi&#10; Redmi Note 6 Pro" value="120" />
                 </map>
               </option>
             </AndroidTestResultsTableState>
