--- app/src/main/java/org/sabaini/thrones/navigation/NavigationManagerImpl.kt
@@ -1,23 +1,0 @@
-package org.sabaini.thrones.navigation
-
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.flow.receiveAsFlow
-import kotlinx.coroutines.launch
-import org.sabaini.thrones.core.di.MainImmediateScope
-import org.sabaini.thrones.core.navigation.NavigationCommand
-import org.sabaini.thrones.core.navigation.NavigationManager
-import javax.inject.Inject
-
-class NavigationManagerImpl @Inject constructor(
-    @MainImmediateScope private val externalMainImmediateScope: CoroutineScope
-) : NavigationManager {
-    private val navigationCommandChannel = Channel<NavigationCommand>(Channel.BUFFERED)
-    override val navigationEvent = navigationCommandChannel.receiveAsFlow()
-
-    override fun navigate(command: NavigationCommand) {
-        externalMainImmediateScope.launch {
-            navigationCommandChannel.send(command)
-        }
-    }
-}
