--- app/src/main/java/org/sabaini/thrones/navigation/NavigationModule.kt
@@ -1,17 +1,0 @@
-package org.sabaini.thrones.navigation
-
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import org.sabaini.thrones.core.navigation.NavigationManager
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-interface NavigationModule {
-
-    @Binds
-    @Singleton
-    fun bindNavigationManager(impl: NavigationManagerImpl): NavigationManager
-}
