+++ core/build.gradle.kts
@@ -1,6 +1,7 @@
 plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.detekt)
+    alias(libs.plugins.hilt)
     alias(libs.plugins.junit)
     alias(libs.plugins.kotlin)
     alias(libs.plugins.kotlin.kapt)
@@ -39,6 +40,7 @@ android {
 
     kotlinOptions {
         freeCompilerArgs = listOf(
+            "-opt-in=androidx.compose.material3.ExperimentalMaterial3Api",
             "-opt-in=kotlinx.coroutines.FlowPreview",
             "-opt-in=kotlinx.coroutines.ExperimentalCoroutinesApi",
             "-opt-in=kotlinx.serialization.ExperimentalSerializationApi"
@@ -53,15 +55,21 @@ android {
 
 dependencies {
     implementation(platform(libs.compose.bom))
-    implementation(libs.bundles.common)
+    implementation(libs.compose.material3)
+    implementation(libs.hilt)
+    implementation(libs.kotlin.coroutines)
     implementation(libs.kotlin.serialization)
     implementation(libs.kotlin.serialization.converter)
     implementation(libs.lifecycle.viewmodel)
+    implementation(libs.navigation)
     implementation(libs.okhttp.logging.interceptor)
     implementation(libs.retrofit)
+    implementation(libs.timber)
     testImplementation(libs.bundles.common.test)
+    androidTestImplementation(libs.bundles.common.android.test)
 
     kapt(libs.hilt.compiler)
+    kaptAndroidTest(libs.test.android.hilt.compiler)
 
     detektPlugins(libs.detekt.twitter.compose)
 }
