+++ core/src/main/java/org/sabaini/thrones/core/navigation/NavigationManager.kt
@@ -1,8 +1,23 @@
 package org.sabaini.thrones.core.navigation
 
-import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.flow.receiveAsFlow
+import kotlinx.coroutines.launch
+import org.sabaini.thrones.core.di.MainImmediateScope
+import javax.inject.Inject
+import javax.inject.Singleton
 
-interface NavigationManager {
-    val navigationEvent: Flow<NavigationCommand>
-    fun navigate(command: NavigationCommand)
+@Singleton
+class NavigationManager @Inject constructor(
+    @MainImmediateScope private val externalMainImmediateScope: CoroutineScope
+) {
+    private val navigationCommandChannel = Channel<NavigationCommand>(Channel.BUFFERED)
+    val navigationEvent = navigationCommandChannel.receiveAsFlow()
+
+    fun navigate(command: NavigationCommand) {
+        externalMainImmediateScope.launch {
+            navigationCommandChannel.send(command)
+        }
+    }
 }
