+++ feature-character/build.gradle.kts
@@ -17,7 +17,7 @@ android {
         minSdk = 24
         targetSdk = 33
 
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner = "org.sabaini.thrones.core.utils.HiltTestRunner"
     }
 
     buildFeatures {
@@ -35,6 +35,8 @@ android {
     }
 
     compileOptions {
+        isCoreLibraryDesugaringEnabled = true
+
         sourceCompatibility = JavaVersion.VERSION_11
         targetCompatibility = JavaVersion.VERSION_11
     }
@@ -53,6 +55,9 @@ android {
     }
 
     sourceSets {
+        getByName("androidTest") {
+            java.srcDir(project(":core").file("src/androidTest/java"))
+        }
         getByName("test") {
             java.srcDir(project(":core").file("src/test/java"))
         }
@@ -63,20 +68,26 @@ dependencies {
     implementation(project(":core"))
 
     implementation(platform(libs.compose.bom))
-    implementation(libs.bundles.common)
     implementation(libs.accompanist.swipe.refresh)
     implementation(libs.coil)
+    implementation(libs.compose.material3)
+    implementation(libs.hilt)
+    implementation(libs.kotlin.coroutines)
+    implementation(libs.navigation)
     implementation(libs.navigation.hilt)
     implementation(libs.kotlin.serialization)
     implementation(libs.retrofit)
     implementation(libs.room)
+    implementation(libs.timber)
     implementation(libs.lifecycle.runtime.compose)
     testImplementation(libs.bundles.common.test)
-    androidTestImplementation(libs.test.android.compose)
-    androidTestImplementation(platform(libs.compose.bom))
+    androidTestImplementation(libs.bundles.common.android.test)
     debugImplementation(libs.debug.compose.manifest)
 
     kapt(libs.hilt.compiler)
+    kaptAndroidTest(libs.test.android.hilt.compiler)
+
+    coreLibraryDesugaring(libs.desugar)
 
     detektPlugins(libs.detekt.twitter.compose)
 }
