+++ feature-character/src/androidTest/java/org/sabaini/thrones/feature/character/data/CharacterDummyData.kt
@@ -1,5 +1,6 @@
 package org.sabaini.thrones.feature.character
 
+import org.sabaini.thrones.feature.character.domain.model.Character
 import org.sabaini.thrones.feature.character.presentation.model.CharacterDisplayable
 
 internal fun generateTestCharactersFromPresentation() = listOf(
@@ -22,6 +23,36 @@ internal fun generateTestCharactersFromPresentation() = listOf(
         imageUrl = ""
     ),
     CharacterDisplayable(
+        id = 2,
+        firstName = "Jon",
+        lastName = "Snow",
+        fullName = "Jon Snow",
+        title = "King of the North",
+        family = "House Stark",
+        imageUrl = ""
+    )
+)
+
+internal fun generateTestCharactersFromDomain() = listOf(
+    Character(
+        id = 0,
+        firstName = "Daenerys",
+        lastName = "Targaryen",
+        fullName = "Daenerys Targaryen",
+        title = "Mother of Dragons",
+        family = "House Targaryen",
+        imageUrl = ""
+    ),
+    Character(
+        id = 1,
+        firstName = "Samwell",
+        lastName = "Tarly",
+        fullName = "Samwell Tarly",
+        title = "Maester",
+        family = "House Tarly",
+        imageUrl = ""
+    ),
+    Character(
         id = 2,
         firstName = "Jon",
         lastName = "Snow",
